[#970] reworked user async thunks the same way as calendar asyncThunks
