+++ common/src/app/store.ts
@@ -2,7 +2,7 @@ import eventsCalendar from '@common/features/Calendars/CalendarSlice'
 import { postMutationRefreshMiddleware } from '@common/features/Calendars/listeners/postMutationRefresh'
 import searchResultReducer from '@common/features/Search/SearchSlice'
 import settingsReducer from '@common/features/Settings/SettingsSlice'
-import userReducer from '@common/features/User/userSlice'
+import userReducer from '@common/features/User/UserSlice'
 import { combineReducers, configureStore } from '@reduxjs/toolkit'
 import loadingReducer from './loadingSlice'
 import { createBrowserHistory } from 'history'
