+++ common/src/features/User/services/index.ts
@@ -1,0 +1,2 @@
+export { getOpenPaasUserDataThunk } from './getOpenPaasUserDataAsync'
+export { updateUserConfigurationsThunk } from './updateUserConfigurationsAsync'
